changed helper functions to a new folder. Refactored video and customer routes to use them
